moved object functions to an separate folder; values() and entries() are now wrappers to Object.values() and Object.entries()
